pyproject.toml
[tool.poetry]
name = "db_vgames"
version = "1.0.0"
description = "Package for my games database client"
authors = ["Cristhian Bravo <green2cerberuz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "=1.3.20"
fastapi = "=0.61.1"
pydantic = "^1.7"
starlette = "=0.13.6"
uvicorn = "=0.12.2"
python-dotenv = "=0.15.0"
psycopg2 = "^2.8.4"
alembic = "=1.4.3"
requests = "=2.24.0"
databases= "=0.4.0"
asyncpg = "=0.21.0"
graphene-sqlalchemy= ">=2.0"

[tool.poetry.dev-dependencies]
pytest = "=6.1.2"
pytest-asyncio = "=0.14.0"
pytest-cov = "=^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=88
skip_glob="*/alembic/versions/*.py"

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "88"
